requirements.txt
PyMuPDF==1.23.8
pdfplumber==0.10.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
colorama==0.4.6
python-dotenv==1.0.0
openai==1.6.1
httpx==0.25.2
chromadb==0.4.18 
